move orchestrator to its own package
