+++ engine/orchestration/src/main/resources/META-INF/cloudstack/core/spring-engine-orchestration-core-context.xml
@@ -51,7 +51,7 @@
     </bean>
 
     <bean id="networkOrchestrator"
-        class="org.apache.cloudstack.engine.orchestration.NetworkOrchestrator">
+        class="org.apache.cloudstack.engine.orchestration.network.NetworkOrchestrator">
         <property name="networkGurus" value="#{networkGurusRegistry.registered}" />
         <property name="networkElements" value="#{networkElementsRegistry.registered}" />
         <property name="ipDeployers" value="#{ipDeployersRegistry.registered}" />
